Squelch a number of warnings

+++ fieldsolver/gridGlue.cpp
@@ -80,7 +80,7 @@ template <typename T, int stencil> void computeCoupling(dccrg::Dccrg<SpatialCell
   }
 
   // Compute where to send data and what to send
-  for(int i=0; i< dccrgCells.size(); i++) {
+  for(uint64_t i=0; i< dccrgCells.size(); i++) {
      //compute to which processes this cell maps
      std::vector<CellID> fsCells = mapDccrgIdToFsGridGlobalID(mpiGrid, dccrgCells[i]);
